requirements.txt
poetry>=2.1.3
pyqt6>=6.9.0
pyqt6-webengine>=6.9.0
sqlalchemy>=2.0.23
alembic>=1.13.1
python-dotenv>=1.0.0
